AB#2659 -- refactoring
